Initial Contract Tests
Some initial contract testing for payment service endpoints, using the Pact framework. Validating the endpoints in the payment service, around getting all payment schedules as well as a new call to get payment schedules for a given claim id. The latter isn't actually used in the app yet, it was added in anticipation of being used and to fill out the contract tests.

Co-Authored-By: Seymourito Pattisson <seymour.pattisson@digital.hmrc.gov.uk>
Co-authored-by: Phil Benson <philbens@gmail.com>